add validation on delete selection schedule

+++ thesis/app/Http/Controllers/SelectionScheduleController.php
@@ -6,17 +6,18 @@ use Illuminate\Http\Request;
 use App\Http\Controllers\Controller;
 use App\Http\Models\Subvocational;
 use App\Http\Models\SelectionSchedule;
+use App\Http\Models\Selection;
 use Auth;
 
 class SelectionScheduleController extends Controller
 {
     /**
-    * Display a listing of the resource.
-    *
-    * @return \Illuminate\Http\Response
-    */
-   public function index(Request $request)
-   {
+     * Display a listing of the resource.
+     * 
+     * @return \Illuminate\Http\Response
+     */
+    public function index(Request $request)
+    {
         $role_id = Auth::user()->roleId();
         
         if ($role_id == 1 || $role_id == 3 || $role_id == 4 || $role_id == 5 || $role_id == 6) {
@@ -27,15 +28,15 @@ class SelectionScheduleController extends Controller
         } else {
             return redirect()->route('registrants.index');
         }
-   }
+    }
 
     /**
-    * Show the form for creating a new resource.
-    *
-    * @return \Illuminate\Http\Response
-    */
-   public function create()
-   {
+     * Show the form for creating a new resource.
+     * 
+     * @return \Illuminate\Http\Response
+     */
+    public function create()
+    {
         $role_id = Auth::user()->roleId();
             
         if ($role_id == 1 || $role_id == 3 || $role_id == 4 || $role_id == 5 || $role_id == 6) {
@@ -45,16 +46,16 @@ class SelectionScheduleController extends Controller
         } else {
             return redirect()->route('registrants.index');
         }
-   }
+    }
 
-   /**
-    * Store a newly created resource in storage.
-    *
-    * @param  \Illuminate\Http\Request  $request
-    * @return \Illuminate\Http\Response
-    */
-    public function store(Request $request)
-    {
+    /**
+     * Store a newly created resource in storage.
+     * 
+     * @param  \Illuminate\Http\Request  $request
+     * @return \Illuminate\Http\Response
+     */
+        public function store(Request $request)
+        {
         $this->validate($request, [
          'sub_vocational_id' => 'required',
          'date' => 'required',
@@ -70,13 +71,13 @@ class SelectionScheduleController extends Controller
     }
  
     /**
-    * Display the specified resource.
-    *
-    * @param  int  $id
-    * @return \Illuminate\Http\Response
-    */
-   public function show($id)
-   {
+     * Display the specified resource.
+     * 
+     * @param  int  $id
+     * @return \Illuminate\Http\Response
+     */
+    public function show($id)
+    {
         $role_id = Auth::user()->roleId();
             
         if ($role_id == 1 || $role_id == 3 || $role_id == 4 || $role_id == 5 || $role_id == 6) {
@@ -86,16 +87,16 @@ class SelectionScheduleController extends Controller
         } else {
             return redirect()->route('registrants.index');
         }
-   }
+    }
 
-   /**
-    * Show the form for editing the specified resource.
-    *
-    * @param  int  $id
-    * @return \Illuminate\Http\Response
-    */
-   public function edit($id)
-   {
+    /**
+     * Show the form for editing the specified resource.
+     * 
+     * @param  int  $id
+     * @return \Illuminate\Http\Response
+     */
+    public function edit($id)
+    {
         $role_id = Auth::user()->roleId();
             
         if ($role_id == 1 || $role_id == 3 || $role_id == 4 || $role_id == 5 || $role_id == 6) {
@@ -107,17 +108,17 @@ class SelectionScheduleController extends Controller
         } else {
             return redirect()->route('registrants.index');
         }
-   }
+    }
 
-   /**
-    * Update the specified resource in storage.
-    *
-    * @param  \Illuminate\Http\Request  $request
-    * @param  int  $id
-    * @return \Illuminate\Http\Response
-    */
-   public function update(Request $request, $id)
-   {
+    /**
+     * Update the specified resource in storage.
+     * 
+     * @param  \Illuminate\Http\Request  $request
+     * @param  int  $id
+     * @return \Illuminate\Http\Response
+     */
+    public function update(Request $request, $id)
+    {
         $this->validate($request, [
             'sub_vocational_id' => 'required',
             'date' => 'required',
@@ -131,19 +132,28 @@ class SelectionScheduleController extends Controller
 
         return redirect()->route('selectionschedules.index')
                         ->with('success','Jadwal seleksi berhasil diedit');
-   }
+    }
 
-   /**
-    * Remove the specified resource from storage.
-    *
-    * @param  int  $id
-    * @return \Illuminate\Http\Response
-    */
-   public function destroy($id)
-   {
-        SelectionSchedule::find($id)->delete();
+    /**
+     * Remove the specified resource from storage.
+     * 
+     * @param  int  $id
+     * @return \Illuminate\Http\Response
+     */
+    public function destroy($id)
+    {
+        $selection = Selection::with('selectionschedule')
+                                ->where('selection_schedule_id', '=', $id)
+                                ->first();
         
-        return redirect()->route('selectionschedules.index')
-                        ->with('success','Jadwal seleksi berhasil dihapus');
-   }
+        if ($selection == null) {
+            SelectionSchedule::find($id)->delete();
+
+            return redirect()->route('selectionschedules.index')
+                             ->with('success','Jadwal seleksi berhasil dihapus');
+        } else {
+            return redirect()->route('selectionschedules.index')
+                             ->with('failed','Jadwal seleksi tidak bisa dihapus karena seleksi sudah dilakukan penilaian');
+        } 
+    }
 }
